code cleanup with custom hooks
