+++ src/pages/authentication/Signup.jsx
@@ -1,64 +1,14 @@
-import { Link, useLocation, useNavigate } from "react-router-dom";
+import { Link } from "react-router-dom";
 import "styles/authentication/Signup.css";
-import { useEffect, useReducer } from "react";
-import axios from "axios";
-import { useAuth } from "contexts";
+import { useEffect } from "react";
+import { useSignup } from "utils";
 
 const Signup = () => {
+  const { signUpDispatch, submitHandler, signUpState } = useSignup();
+
   useEffect(() => {
     document.title = "Signup | Shopzila";
   }, []);
-
-  const {setIsLoggedIn} = useAuth();
-  const location = useLocation();
-  const navigate = useNavigate();
-
-  const signUpReducer = (signUpState, signUpAction) => {
-    switch (signUpAction.type) {
-      case "FIRSTNAME":
-        return { ...signUpState, firstName: signUpAction.payload.target.value };
-      case "LASTNAME":
-        return { ...signUpState, lastName: signUpAction.payload.target.value };
-      case "EMAIL":
-        return { ...signUpState, email: signUpAction.payload.target.value };
-      case "PASSWORD":
-        return { ...signUpState, password: signUpAction.payload.target.value };
-      case "CONFIRM_PASSWORD":
-        return {
-          ...signUpState,
-          confirmPassword: signUpAction.payload.target.value,
-        };
-      default:
-        return signUpState;
-    }
-  };
-
-  const [signUpState, signUpDispatch] = useReducer(signUpReducer, {
-    firstName: "",
-    lastName: "",
-    email: "",
-    password: "",
-    confirmPassword: "",
-  });
-
-  const submitHandler = async(event) => {
-    event.preventDefault();
-    try{
-      const response = await axios.post(`/api/auth/signup`, {
-        firstName: signUpState.firstName,
-        lastName: signUpState.lastName,
-        email: signUpState.email,
-        password: signUpState.password,
-        confirmPassword: signUpState.confirmPassword,
-      });
-      localStorage.setItem("token", response.data.encodedToken);
-      setIsLoggedIn(true);
-      navigate(location.state?.from?.pathname || "/product", { replace: true });
-    }
-    catch(e){
-      console.log(e);
-    }
-  };
 
   return (
     <section className="signupContainer">
