+++ src/reducer/index.js
@@ -1,3 +1,4 @@
 export { cartReducer } from "./CartReducer";
 export { reducer } from "./ProductReducer";
 export { wishlistReducer } from "./WishlistReducer";
+export { signUpReducer } from "./Signup";
