+++ src/utils/index.js
@@ -1,3 +1,15 @@
-export { addToCart, removeFromCart, increaseQuantity, decreaseQuantity } from "./cart";
-export { getCategorisedProduct, getRatingProducts, getSortingProduts } from "./filter";
+export {
+  addToCart,
+  removeFromCart,
+  increaseQuantity,
+  decreaseQuantity,
+} from "./cart";
+export {
+  getCategorisedProduct,
+  getRatingProducts,
+  getSortingProduts,
+} from "./filter";
 export { addToWishlist, removeFromWishlist } from "./wishlist";
+export { useLogin } from "./login";
+export { useLogout } from "./logout";
+export { useSignup } from "./signup";
